try
sd

+++ Assets/Scripts/Droids.cs
@@ -14,7 +14,7 @@ public class Droids : MonoBehaviour
     [SerializeField] private float laserSpeed = 10f;
     [SerializeField] private float returnRotationSpeed = 2f;
     [SerializeField] private float helth = 100f;
-    [SerializeField] private float DroidDamage = 10f;
+    [SerializeField] private int DroidDamage = 10;
     [SerializeField] private float critModifier = 2f;
     [SerializeField] private ParticleSystem explosionEffect;
 
@@ -97,6 +97,8 @@ public class Droids : MonoBehaviour
         GameObject laser = Instantiate(laserPrefab, shootPoint.position, shootPoint.rotation);
         laser.GetComponent<Laser>().setDirection(shootPoint.forward, laserSpeed, DroidDamage);
         hasShot = false;
+
+
     }
 
     public void TakeDamage(float damage, bool isCrit)
